Installation section, plus an interface change, plus a plugin
Notes:
I added "admonitions" to the site so we can do big WARNING boxes on the docs, etc.
site_url is supposed to be required in MkDocs. I think we may be able to have a link in the docs back to the Classic website, which is something I'm very much for.

I added a warning on the system requirements page about plugin dependencies, using one example I was aware of. Will add more, ideally all of them, over time.
Because I mention PdfText in the sysreqs page, I edited the plugin page too.  I will do more when I get to that section properly (updating images, etc.).
Sysreqs now has the upgrade-your-dependencies info from Installation - made more sense there I think.

Setting Directory Permissions should be expanded more - with examples, screenshots of where to look for permissions, etc. - just did a short intro to what the heck we're talking about for now. That or we find a reliable external resource to send people to.

Updated the language page, but cutoff at 5% interface translation. Should be higher perhaps? Maybe as much as 50% or more?

Installation: I cut "LAMP server setup". Unless there is an alternative process, this is THE installation method, it doesn't need a name.
Clarified some one-click install stuff. More soon, but mentioning it here makes sense (not Dreamhost anymore).
Removed some deprecated language here about there being 2 .htaccess files? Downloaded the 1.3.1 package to be sure how that worked. Happy to be corrected if what I've written is inaccurate.
Clarified that you have to recompress your updated files and then upload that, if you have command line access.

Do we want to institute a cutoff for mentions of older versions? Like anything before 2.0 doesn't need to be mentioned in the docs? Or we could fork this doc for a new 3.+ version and cut all the old mentions out of that.